requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-tenants==3.6.1
psycopg2-binary